+++ tessellate/utils/__init__.py
@@ -1,0 +1,6 @@
+
+
+__author__ = """Christopher Bevan Barnett"""
+__email__ = 'chrisbarnettster@gmail.com'
+__version__ = '0.3.0'
+