+++ HISTORY.rst
@@ -2,7 +2,16 @@
 History
 =======
 
+0.3.0 (2017-11-23)
+------------------
+* First release on PyPi
+
+0.2.0 (2017-11-23)
+------------------
+* Alpha version that can read PDBlists and builtin, can write json
+* Include examples and much verbose logging
+
 0.1.0 (2017-11-21)
 ------------------
 
-* First release on PyPI.
+* Alpha version. Basic function